+++ backend/src/database/migrations/index.ts
@@ -10,6 +10,7 @@ import { V050A1672154027079 } from './1672154027079-V_0_5_0_a.js';
 import { V060A1790293063219 } from './1790293063219-V_0_6_0_a.js';
 import { V060B1790297214057 } from './1790297214057-V_0_6_0_b.js';
 import { V060C1790333466919 } from './1790333466919-V_0_6_0_c.js';
+import { V060D1790335094391 } from './1790335094391-V_0_6_0_d.js';
 
 export const MigrationList: Newable<MigrationInterface>[] = [
   V030A1661692206479,
@@ -22,4 +23,5 @@ export const MigrationList: Newable<MigrationInterface>[] = [
   V060A1790293063219,
   V060B1790297214057,
   V060C1790333466919,
+  V060D1790335094391,
 ];