+++ backend/src/routes/api/api.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ApiKeysModule } from './apikeys/apikeys.module.js';
+import { GalleryApiModule } from './gallery/gallery.module.js';
 import { InfoModule } from './info/info.module.js';
 import { PrefModule } from './pref/pref.module.js';
 import { RolesApiModule } from './roles/roles.module.js';
@@ -14,6 +15,7 @@ import { UserApiModule } from './user/user.module.js';
     RolesApiModule,
     ApiKeysModule,
     UsageApiModule,
+    GalleryApiModule,
   ],
 })
 export class PicsurApiModule {}